// packages/shared/src/enums/AreaType.ts
export enum AreaType {
  /** 台北 */
  TAIPEI = 'Taipei',
  /** 桃園市 */
  TAOYUAN = 'Taoyuan',
  /** 台中 */
  TAICHUNG = 'Taichung',
  /** 台南 */
  TAINAN = 'Tainan',
  /** 高雄 */
  KAOHSIUNG = 'Kaohsiung',
  /** 基隆 */
  KEELUNG = 'Keelung',
  /** 新竹 */
  HSINCHU = 'Hsinchu',
  /** 苗栗 */
  MIAOLI = 'Miaoli',
  /** 彰化 */
  CHANGHUA = 'Changhua',
  /** 南投 */
  NANTOU = 'Nantou',
  /** 雲林 */
  YUNLIN = 'Yunlin',
  /** 嘉義 */
  CHIAYI = 'Chiayi',
  /** 屏東 */
  PINGTUNG = 'Pingtung',
  /** 宜蘭 */
  YILAN = 'Yilan',
  /** 花蓮縣 */
  HUALIEN = 'Hualien',
  /** 台東縣 */
  TAITUNG = 'Taitung',
  /** 金門縣 */
  KINMEN = 'Kinmen',
  /** 澎湖縣 */
  PENGHU = 'Penghu',
  /** 連江 */
  LIENCHIANG = 'Lienchiang',
}