// packages/shared/src/enums/CityNameType.ts
export enum CityNameType {
  /** 台北市 */
  TAIPEI = 'Taipei',
  /** 新北市 */
  NEW_TAIPEI = 'NewTaipei',
  /** 桃園市 */
  TAOYUAN = 'Taoyuan',
  /** 台中市 */
  TAICHUNG = 'Taichung',
  /** 台南市 */
  TAINAN = 'Tainan',
  /** 高雄市 */
  KAOHSIUNG = 'Kaohsiung',
  /** 基隆市 */
  KEELUNG = 'Keelung',
  /** 新竹市 */
  HSINCHU = 'Hsinchu',
  /** 新竹縣 */
  HSINCHU_COUNTY = 'HsinchuCounty',
  /** 苗栗縣 */
  MIAOLI_COUNTY = 'MiaoliCounty',
  /** 彰化縣 */
  CHANGHUA_COUNTY = 'ChanghuaCounty',
  /** 南投縣 */
  NANTOU_COUNTY = 'NantouCounty',
  /** 雲林縣 */
  YUNLIN_COUNTY = 'YunlinCounty',
  /** 嘉義縣 */
  CHIAYI_COUNTY = 'ChiayiCounty',
  /** 嘉義市 */
  CHIAYI = 'Chiayi',
  /** 屏東縣 */
  PINGTUNG_COUNTY = 'PingtungCounty',
  /** 宜蘭縣 */
  YILAN_COUNTY = 'YilanCounty',
  /** 花蓮縣 */
  HUALIEN_COUNTY = 'HualienCounty',
  /** 台東縣 */
  TAITUNG_COUNTY = 'TaitungCounty',
  /** 金門縣 */
  KINMEN_COUNTY = 'KinmenCounty',
  /** 澎湖縣 */
  PENGHU_COUNTY = 'PenghuCounty',
  /** 連江縣（馬祖） */
  LIENCHIANG_COUNTY = 'LienchiangCounty',
}